refactor: 학생관련 dto 파일 경로 변경
