+++ src/test/java/com/time/studentmanage/service/StudentServiceTest.java
@@ -1,9 +1,9 @@
 package com.time.studentmanage.service;
 
 import com.time.studentmanage.domain.member.Student;
-import com.time.studentmanage.dto.StudentRespDto;
-import com.time.studentmanage.dto.StudentSaveReqDto;
-import com.time.studentmanage.dto.StudentUpdateReqDto;
+import com.time.studentmanage.domain.dto.student.StudentRespDto;
+import com.time.studentmanage.domain.dto.student.StudentSaveReqDto;
+import com.time.studentmanage.domain.dto.student.StudentUpdateReqDto;
 import com.time.studentmanage.repository.StudentRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
